third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-font-in-main-frame-expected.txt

CONSOLE WARNING: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-font.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/resources/Ahem.ttf': this content should also be loaded over HTTPS.

CONSOLE WARNING: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-font.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/resources/Ahem.ttf': this content should also be loaded over HTTPS.

CONSOLE WARNING: Blink is considering rejecting non spec-compliant cross-origin web font requests: http://127.0.0.1:8080/resources/Ahem.ttf. Please use Access-Control-Allow-Origin to make these requests spec-compliant.
CONSOLE WARNING: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-font.html' was loaded over HTTPS, but ran insecure content from 'http://127.0.0.1:8080/resources/Ahem.ttf': this content should also be loaded over HTTPS.

This test opens a window that loads an insecure font. We should trigger a mixed content callback because the main frame in the window is HTTPS but is displaying insecure content.
